Refactor task creation
